Implement concern pattern for problem logic
